final_project/trek_predict/requirements.txt
ansible==1.7.2
apache-log-parser==1.7.0
archmage==0.2.4
argcomplete==0.8.1
argh==0.25.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
BeautifulSoup==3.2.1
carbon==0.9.12
CDDB==1.4
certifi==2015.11.20
cffi==0.8.6
characteristic==0.1.0
chardet==2.3.0
chm2pdf==0.9
colorama==0.3.2
Couchapp==1.0.1
CouchDB==0.10
cryptography==0.6.1
cssselect==0.9.1
decorator==4.0.6
defusedxml==0.4.1
dh-virtualenv==0.6.99
Django==1.7.7
django-tagging==0.3.1
dnspython==1.12.0
docutils==0.12
ecdsa==0.11
feedparser==5.1.3
FormEncode==1.3.0
gevent==1.0.2
gpodder==3.8.1
graphite-web==0.9.12
greenlet==0.4.9
gyp==0.1
html5lib==0.999
http-parser==0.8.3
httplib2==0.9
hulu==0.1.0
IMDbPY==5.0
ipython==2.3.0
ipython-genutils==0.1.0
Jinja2==2.7.3
jsmin==2.0.9
jupyter-core==4.0.6
keyring==4.0
lxml==3.5.0
MarkupSafe==0.23
mutagen==1.25.1
mygpoclient==1.7
ndg-httpsclient==0.3.2
numpy==1.8.2
oauthlib==1.0.3
PAM==0.4.2
pandas==0.17.1
paramiko==1.15.1
path.py==8.1.2
pathtools==0.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.5
Pillow==2.6.1
pithos==0.3
ply==3.4
ptyprocess==0.5
pyasn1==0.1.7
pyasn1-modules==0.0.5
pychm==0.8.4.1
pycparser==2.10
pycrypto==2.6.1
pycups==1.9.63
pycurl==7.19.5
pyflix2==0.2.0
Pygments==2.0.1
pygobject==3.14.0
pyinotify==0.9.4
pynzb==0.1.0
pyOpenSSL==0.14
pyparsing==2.0.3
pyquery==1.2.4
pyserial==2.6
pysmbc==1.0.15.3
python-apt==0.9.3.12
python-dateutil==2.4.2
python-debian==0.1.27
python-debianbts==1.11
python-keyczar==0.71rc0
pytz==2012rc0
pyxdg==0.25
PyYAML==3.11
pyzmq==14.4.0
quodlibet==3.2.2
reportbug==6.6.3
reportlab==3.1.8
requests==2.9.0
requests-oauthlib==0.6.0
restkit==4.2.2
roman==2.0.0
rottentomatoes==2.1
SecretStorage==2.1.1
service-identity==1.0.0
simplegeneric==0.8.1
simplejson==3.6.5
singledispatch==3.4.0.3
six==1.10.0
SOAPpy==0.12.22
socketpool==0.5.3
SQLAlchemy==1.0.10
sqlalchemy-migrate==0.10.0
SQLObject==2.1.2
sqlparse==0.1.18
stevedore==0.14.1
Tempita==0.5.2
tornado==4.3
traitlets==4.0.0
Twisted-Core==14.0.2
ua-parser==0.5.0
urllib3==1.9.1
user-agents==1.0.1
uTidylib==0.2
uWSGI==2.0.11.2
vboxapi==1.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
watchdog==0.7.0
WebOb==1.4
wheel==0.24.0
whisper==0.9.12
wstools==0.4.3
wxPython==3.0.1.1
wxPython-common==3.0.1.1
zim==0.62
zope.interface==4.1.1
